% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{bm_family_name}
\alias{bm_family_name}
\title{Extract a family name}
\usage{
bm_family_name(family)
}
\arguments{
\item{family}{A family object or family constructor.}
}
\value{
A single family name string when available.
}
\description{
Extract a family name
}
\keyword{internal}
