% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{bm_is_ppml_family}
\alias{bm_is_ppml_family}
\title{Detect PPML-style families}
\usage{
bm_is_ppml_family(family)
}
\arguments{
\item{family}{A family object or family constructor.}
}
\value{
\code{TRUE} for \code{ppml()}-style families, otherwise \code{FALSE}.
}
\description{
Detect PPML-style families
}
\keyword{internal}
