% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{bm_one_index}
\alias{bm_one_index}
\title{Convert cluster labels to one-based indexing}
\usage{
bm_one_index(x)
}
\arguments{
\item{x}{An integer vector or list of integer vectors.}
}
\value{
\code{x} with labels shifted up by one.
}
\description{
Convert cluster labels to one-based indexing
}
\keyword{internal}
