% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{bm_partition_equal}
\alias{bm_partition_equal}
\title{Compare partitions while ignoring label permutations}
\usage{
bm_partition_equal(a, clu2)
}
\arguments{
\item{a, clu2}{Vector or list partitions to compare.}
}
\value{
\code{TRUE} when the partitions are equivalent up to relabeling.
}
\description{
Compare partitions while ignoring label permutations
}
\keyword{internal}
