% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{bm_zero_index}
\alias{bm_zero_index}
\title{Convert cluster labels to zero-based indexing}
\usage{
bm_zero_index(x)
}
\arguments{
\item{x}{An integer vector or list of integer vectors.}
}
\value{
\code{x} with labels shifted down by one.
}
\description{
Convert cluster labels to zero-based indexing
}
\keyword{internal}
