% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsbm.R
\name{mdsbm_icl_one_partition}
\alias{mdsbm_icl_one_partition}
\title{Compute ICL for one supplied dynamic blockmodeling partition}
\usage{
mdsbm_icl_one_partition(
  x,
  sets,
  k,
  clu,
  symmetric = TRUE,
  epsilon = 0.001,
  epsilonTrans = 0.001
)
}
\arguments{
\item{x}{Square adjacency matrix.}

\item{sets}{Integer vector of time-set sizes.}

\item{k}{Integer vector with the number of ordinary clusters per set.}

\item{clu}{List of zero-based cluster membership vectors, one per set.}

\item{symmetric}{Whether to use the symmetric likelihood.}

\item{epsilon}{Lower/upper clipping value for intra-time densities.}

\item{epsilonTrans}{Lower/upper clipping value for inter-time transition
probabilities.}
}
\description{
Compute ICL for one supplied dynamic blockmodeling partition
}
