% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_optimizer.R
\name{print.dynamic_glm_blockmodel}
\alias{print.dynamic_glm_blockmodel}
\title{Print a dynamic GLM-Markov optimizer fit}
\usage{
\method{print}{dynamic_glm_blockmodel}(x, ...)
}
\arguments{
\item{x}{A \code{dynamic_glm_blockmodel} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dynamic GLM-Markov optimizer fit
}
