% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_network.R
\name{print.dynamic_network}
\alias{print.dynamic_network}
\title{Print a dynamic network object}
\usage{
\method{print}{dynamic_network}(x, ...)
}
\arguments{
\item{x}{A \code{dynamic_network} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dynamic network object
}
