% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_glm.R
\name{print.time_glm_blockmodel}
\alias{print.time_glm_blockmodel}
\title{Print a time-specific GLM blockmodel fit}
\usage{
\method{print}{time_glm_blockmodel}(x, ...)
}
\arguments{
\item{x}{A \code{time_glm_blockmodel} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a time-specific GLM blockmodel fit
}
