% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_glm.R
\name{print.time_glm_blockmodels}
\alias{print.time_glm_blockmodels}
\title{Print a multi-time GLM blockmodel fit}
\usage{
\method{print}{time_glm_blockmodels}(x, ...)
}
\arguments{
\item{x}{A \code{time_glm_blockmodels} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a multi-time GLM blockmodel fit
}
