% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_network.R
\name{summary.dynamic_network}
\alias{summary.dynamic_network}
\title{Summarize a dynamic network object}
\usage{
\method{summary}{dynamic_network}(object, ...)
}
\arguments{
\item{object}{A \code{dynamic_network} object.}

\item{...}{Ignored.}
}
\value{
A list with compact summary information.
}
\description{
Summarize a dynamic network object
}
