% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_glm.R
\name{summary.time_glm_blockmodel}
\alias{summary.time_glm_blockmodel}
\title{Summarize a time-specific GLM blockmodel fit}
\usage{
\method{summary}{time_glm_blockmodel}(object, ...)
}
\arguments{
\item{object}{A \code{time_glm_blockmodel} object.}

\item{...}{Ignored.}
}
\value{
A compact summary list.
}
\description{
Summarize a time-specific GLM blockmodel fit
}
