% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_glm.R
\name{summary.time_glm_blockmodels}
\alias{summary.time_glm_blockmodels}
\title{Summarize a multi-time GLM blockmodel fit}
\usage{
\method{summary}{time_glm_blockmodels}(object, ...)
}
\arguments{
\item{object}{A \code{time_glm_blockmodels} object.}

\item{...}{Ignored.}
}
\value{
A compact summary list.
}
\description{
Summarize a multi-time GLM blockmodel fit
}
